changed the structure of the project
