+++ src/components/serverCharts/login-attepts-chart.jsx
@@ -1,51 +1,34 @@
 import { ResponsivePie } from "@nivo/pie";
-import loginAttemptChartService from "../services/loginAttemptChartService";
 import { useEffect, useState } from "react";
+import serverDataService from "../../services/serverDataService";
 
 const LoginAttemptsChart = () => {
   const [data, setData] = useState([]);
 
-  let loginAttempts = {
-    failed: 0,
-    successed: 0,
-  };
-
   useEffect(() => {
-    loginAttemptChartService.getData().then((response) => {
-      setData(response);
+    serverDataService.getLoginAttempts().then((res) => {
+      setData(res);
     });
   }, []);
 
-  loginAttempts = data.reduce((y, x) => {
-    if (x.status === "success") {
-      y.successed++;
-    }
-    if (x.status === "failed") {
-      y.failed++;
-    }
-    return y;
-  }, loginAttempts);
-
-  console.log(loginAttempts);
-  
   const chartData = [
     {
       id: "failed",
       label: "epäonnistuneet",
-      value: loginAttempts.failed,
+      value: data.failed,
       color: "hsl(90, 70%, 50%)",
     },
     {
       id: "successed",
       label: "onnistuneet",
-      value: loginAttempts.successed,
+      value: data.successed,
       color: "hsl(56, 70%, 50%)",
     },
   ];
 
   return (
     <div className="chart-container">
-    <h3>Login attempts</h3>
+      <h3>Login attempts</h3>
       <ResponsivePie
         data={chartData}
         margin={{ top: 40, right: 80, bottom: 80, left: 80 }}
