+++ src/components/serverCharts/serverDataComponent.jsx
@@ -1,0 +1,11 @@
+import LoginAttemptsChart from "./login-attepts-chart";
+
+const ServerDataComponent = () => {
+  return (
+    <>
+      <LoginAttemptsChart />
+    </>
+  );
+};
+
+export default ServerDataComponent;
